# improve parallelism to chunk reader of binary data