--- a/docs/notebooks/Authentication.ipynb
+++ b/docs/notebooks/Authentication.ipynb
@@ -62,8 +62,8 @@
    "source": [
     "from pydap.cas.urs import setup_session\n",
     "\n",
-    "username = \"<YourUserName>\" # Earthdata URS Login Get account here https://urs.earthdata.nasa.gov/\n",
-    "password = \"<YourPassword>\"\n",
+    "username = \"mikejmnez\" # Earthdata URS Login Get account here https://urs.earthdata.nasa.gov/\n",
+    "password = \"FHD!xua*zqy6mkm6fgk\"\n",
     "\n",
     "hyrax_url = 'https://opendap.earthdata.nasa.gov/collections/C2532426483-ORNL_CLOUD/granules/'\n",
     "session = setup_session(username, password, check_url=hyrax_url)"
@@ -135,7 +135,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "edl_token = \"<YourToken>\"\n",
+    "edl_token = \"eyJ0eXAiOiJKV1QiLCJvcmlnaW4iOiJFYXJ0aGRhdGEgTG9naW4iLCJzaWciOiJlZGxqd3RwdWJrZXlfb3BzIiwiYWxnIjoiUlMyNTYifQ.eyJ0eXBlIjoiVXNlciIsInVpZCI6Im1pa2VqbW5leiIsImV4cCI6MTczNzM5NTM1NCwiaWF0IjoxNzMyMjExMzU0LCJpc3MiOiJodHRwczovL3Vycy5lYXJ0aGRhdGEubmFzYS5nb3YiLCJpZGVudGl0eV9wcm92aWRlciI6ImVkbF9vcHMiLCJhc3N1cmFuY2VfbGV2ZWwiOjJ9.Xc9IOvVT4myF3FAM2o88yOln0j6MyM39u4NskuoIXSWhHesFd4eQRAku-zbeghgFZgnjzavyUqQHtRhNbLE1IQpyisxKjpcs-DhUfsQsj3c5gxYWxwm4YQjJgbJp7mja327rtulcQGBJlxV_r05YjuituDyq376pkN-GgEZyAzL2e5RIgb0NUEOWwrjVYGA379YrEnf_v3SK4OO-Ur_L04XQGyW82zfVwgrYLlCq2HyUiafMfwTRK3OhhXRzbVHa5t4YNIW7t4m2V7qxXmj9eGTxHtYwsJFJyfwOSl64UKlWzSEjbUkd52GU6xqv9NCa1a1_9UiUZcYNUxGx6m0mwA\"\n",
     "\n",
     "auth_hdr=\"Bearer \" + edl_token\n",
     "\n",