--- a/docs/notebooks/CMIP6.ipynb
+++ b/docs/notebooks/CMIP6.ipynb
@@ -345,7 +345,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.12.4"
+   "version": "3.11.10"
   }
  },
  "nbformat": 4,