--- a/src/pydap/client.py
+++ b/src/pydap/client.py
@@ -148,7 +148,7 @@ def open_dap_file(file_path, das_path=None):
         # if f.peek()[0:2] == b'\x04\x00':
         f.seek(dmr_len)
         _ = f.read(4)
-        pydap.handlers.dap.unpack_dap4_data(f, dataset)
+        pydap.handlers.dap.unpack_dap4_data(f.read(), dataset)
     return dataset