… docs/5_minute_tutorial.ipynb
… docs/notebooks/ECCO.ipynb
… docs/notebooks/PACE.ipynb
… docs/notebooks/SWOT.ipynb