<!-- Thank you very much for your hard work and contribution! -->

# ~Draft~ Ready for review
The following Pull Request:

- [x] Closes #418 by improving performance (3 orders of mangnitude) when numpy-ficating binary data from the dap response.
- [x] ~Some tests are failing, and I need to write a test or two to make sure things work~ Tests are passing :)



## benchmark

* download .dap response to a single variable from hyrax
```python
import numpy
from pydap.handlers.dap import safe_dmr_and_data, dmr_to_dataset, unpack_dap4_data, BytesReader, stream2bytearray, newstream2bytearray
from pydap.net import GET

# prep code
url = "http://oceandata.sci.gsfc.nasa.gov/opendap/PACE_OCI/L3SMI/2024/1203/PACE_OCI.20241203.L3m.DAY.CHL.V2_0.chlor_a.4km.NRT.nc.dap"
r = GET(url+"?dap4.ce=/chlor_a")

dmr, data = safe_dmr_and_data(r, 'ascii', url)
dataset = dmr_to_dataset(dmr)
xdr_stream = BytesReader(data)

```

### tests:

**before PR**
```python
%%time
buffer = stream2bytearray(xdr_stream)
>> CPU times: user 1min 3s, sys: 1min 46s, total: 2min 49s
Wall time: 2min 50s
```

**New**
```python
%%time
newbuffer = stream2bytearray(data)
>>> CPU times: user 199 ms, sys: 73.5 ms, total: 272 ms
Wall time: 266 ms
```