@Mikejmnez 2024-12-06
@jgallagher59701 this is ready for review :).

Speed up of close to 3 orders of magnitude, when numpy-fying binary data from dap responses (3 orders of magnitude speed uup is an extreme case - in most cases probebly 1 order of magnitude)