@Mikejmnez 2024-12-06
merging so this is installable and we can update the documentation